ms-agent-harness/migrated_code/order-processor/requirements.txt
azure-functions
azure-cosmos
azure-storage-blob
azure-storage-queue